+++ src/red/team/boss/list_boss/BLACK/ZamasMax.java
@@ -23,8 +23,8 @@ public class ZamasMax extends Boss {
 
     @Override
     public void reward(Player plKill) {
-        byte randomDo = (byte) new Random().nextInt(Manager.itemIds_TL.length - 1);
-        byte randomNR = (byte) new Random().nextInt(Manager.itemIds_NR_SB.length -1);
+        short randomDo = (short) new Random().nextInt(Manager.itemIds_TL.length - 1);
+        short randomNR = (short) new Random().nextInt(Manager.itemIds_NR_SB.length -1);
 
         if (Util.isTrue(10, 20)) {
 //                int[] manhthuong = new int[] { 1066, 1067, 1068, 1069, 1070 };  
@@ -37,7 +37,7 @@ public class ZamasMax extends Boss {
 //                    Service.gI().dropItemMap(this.zone,
 //                            Util.manhTS(zone, manhthuong[randomAWJ], 1, this.location.x, this.location.y, plKill.id));
                 Service.gI().dropItemMap(this.zone,
-                        Util.ratiItem(zone, (byte) 15, 1, this.location.x + 2, this.location.y, plKill.id));
+                        Util.ratiItem(zone, (short) 15, 1, this.location.x + 2, this.location.y, plKill.id));
             } else {
                 Service.gI().dropItemMap(this.zone,
                         Util.manhTS(zone, manhhiem[randomGR], 3, this.location.x, this.location.y, plKill.id));
